+++ src/index.ts
@@ -1,5 +1,7 @@
 export * as sm2 from "./sm2/index";
 export { SM3 } from "./sm2/sm3";
+export { kdf } from './sm2/kdf';
+export { hkdf } from './sm3/hkdf';
 export { sm3 } from "./sm3/index";
 export * as sm4 from "./sm4/index";
 export { hexToArray } from "./sm2/utils";