+++ src/sm2/kx.ts
@@ -1,43 +1,12 @@
 import { field, sm2Curve } from './ec';
 import { KeyPair, hexToArray, leftPad } from './utils';
 import * as utils from '@noble/curves/abstract/utils';
-import { sm3 } from './sm3';
 import { EmptyArray, getZ } from '.';
-
+import { kdf } from './kdf';
 
 // 用到的常数
 const wPow2 = utils.hexToNumber('80000000000000000000000000000000')
 const wPow2Sub1 = utils.hexToNumber('7fffffffffffffffffffffffffffffff')
-
-// from sm2 sign part, extracted for code reusable.
-function hkdf(z: Uint8Array, keylen: number) {
-  let msg = new Uint8Array(keylen)
-  let ct = 1
-  let offset = 0
-  let t = EmptyArray
-  const ctShift = new Uint8Array(4)
-  const nextT = () => {
-    // (1) Hai = hash(z || ct)
-    // (2) ct++
-    ctShift[0] = ct >> 24 & 0x00ff
-    ctShift[1] = ct >> 16 & 0x00ff
-    ctShift[2] = ct >> 8 & 0x00ff
-    ctShift[3] = ct & 0x00ff
-    t = sm3(utils.concatBytes(z, ctShift))
-    ct++
-    offset = 0
-  }
-  nextT() // 先生成 Ha1
-
-  for (let i = 0, len = msg.length; i < len; i++) {
-    // t = Ha1 || Ha2 || Ha3 || Ha4
-    if (offset === t.length) nextT()
-
-    // 输出 stream
-    msg[i] = t[offset++] & 0xff
-  }
-  return msg
-}
 
 export function calculateSharedKey(
   keypairA: KeyPair,
@@ -78,6 +47,6 @@ export function calculateSharedKey(
   // KA = KDF(xU || yU || ZA || ZB, kLen)
   const xU = hexToArray(leftPad(utils.numberToHexUnpadded(U.x), 64))
   const yU = hexToArray(leftPad(utils.numberToHexUnpadded(U.y), 64))
-  const KA = hkdf(utils.concatBytes(xU, yU, ZA, ZB), sharedKeyLength)
+  const KA = kdf(utils.concatBytes(xU, yU, ZA, ZB), sharedKeyLength)
   return KA
 }